code/user/requirements.txt
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
Flask==1.1.2
idna==2.10
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.5.3
requests==2.24.0
simplejson==3.17.2
six==1.15.0
toml==0.10.1
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
PyJWT==1.7.1
prometheus-flask-exporter==0.18.1